+++ web/api/chat_completion_handler.py
@@ -9,7 +9,7 @@ from docq.model_selection.main import get_model_settings_collection
 from pydantic import Field, ValidationError
 from tornado.web import HTTPError
 
-from web.api.base import BaseRequestHandler
+from web.api.base_handlers import BaseRequestHandler
 from web.api.models import MessageModel, MessageResponseModel
 from web.api.utils import CamelModel, authenticated
 from web.utils.streamlit_application import st_app
