+++ web/api/space_handler.py
@@ -7,7 +7,7 @@ from docq.data_source.list import SpaceDataSources
 from pydantic import BaseModel, ValidationError
 from tornado.web import HTTPError
 
-from web.api.base import BaseRagRequestHandler
+from web.api.base_handlers import BaseRagRequestHandler
 from web.api.utils import authenticated
 from web.utils.streamlit_application import st_app
