+++ web/api/thread_handler.py
@@ -6,7 +6,7 @@ from docq.domain import FeatureKey, OrganisationFeatureType
 from pydantic import ValidationError
 from tornado.web import HTTPError
 
-from web.api.base import BaseRequestHandler
+from web.api.base_handlers import BaseRequestHandler
 from web.api.models import ThreadModel, ThreadPostRequestModel, ThreadResponseModel
 from web.api.utils import authenticated
 from web.utils.streamlit_application import st_app
